requirements.txt
discord.py
langchain
openai
tiktoken
python-dotenv